requirements.txt
fastapi[all]
uvicorn[standard]

strawberry-graphql
sqlalchemy
asyncpg
